COMP2001_CW2_API/requirements.txt
Flask==2.2.2
connexion[swagger-ui]==2.14.1
Werkzeug==2.2.2

pyodbc==5.1.0
python-dotenv==1.0.1

requests==2.31.0

